requirements.txt
pybit~=1.3.6
requests~=2.25.1
peewee~=3.14.9
binance-futures-connector
environs~=9.5.0
discord~=1.7.3
loguru~=0.6.0
pymysql
discord.py
cryptography
